Move abstract Bee to game/characters/base
