+++ frontend/models/game/characters/Drone.php
@@ -9,7 +9,7 @@
 namespace frontend\models\game\characters;
 
 
-use frontend\models\game\base\Bee;
+use frontend\models\game\characters\base\Bee;
 use frontend\models\game\base\CharacterInterface;
 use frontend\models\game\characters\interfaces\DroneInterface;
